Implement abstract class method callbacks.
Perl uses identical syntax for class and object method invocation.
The only difference is the invocant---either a class or an object.
Implemented class method callbacks in terms of object method callbacks
for now.  The implementations may diverge in the future if the
semantics need to, but the interfaces should be forward compatible.